requirements.txt
dash
pandas
datetime
sqlalchemy
requests
dash_bootstrap_components
dash_auth
plotly
plotly-express
dash-table
psycopg2-binary
dash-core-components 
dash-html-components
dash_auth
flask
authlib
six
python-dotenv
dash-gif-component